pyproject.toml
[tool.poetry]
name = "pyPDFChat"
version = "0.0.1"
description = ""
authors = ["Your Name <34923485+Haste171@users.noreply.github.com>"]
license = "MIT"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
# langchain-pinecone richiede Python >=3.9,<3.13
python = ">=3.9,<3.13"

python-dotenv = "*"
pinecone-client = "*"
colorama = "*"
streamlit = "*"
watchdog = "*"
python-multipart = "*"
pymupdf = "*"
fastapi = "*"
uvicorn = "*"
docx2txt = "*"
langchain = "*"
langchain-openai = "*"
langchain-anthropic = "*"
langchain-community = "*"
langchain-pinecone = "*"
langchain-chroma = "*"
altair = "4.2.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
